Extract and handle jobs in one function
