+++ utils/__init__.py
@@ -1,2 +1,1 @@
-from feed_parsing import extract_jobs
-from handle_jobs import handle_new_jobs
+from feed_parsing import extract_and_handle_jobs
